Refactor Contribution components
